requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
alchemy-mock==0.4.3
aniso8601==9.0.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.1.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
MarkupSafe==2.1.1
mock==4.0.3
numpy==1.23.4
packaging==22.0
pandas==1.5.1
pipfile==0.0.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.45
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.1.2
zipp==3.11.0
